+++ vector.h
@@ -12,7 +12,7 @@ public:
    inline void clear() { m_rg.clear(); }
    inline int size() const { return (int)m_rg.size(); }
    inline int capacity() const { return (int)m_rg.capacity(); }
-   inline void resize(const int i) { m_rg.resize(i); }
+   inline void resize(const size_t i) { m_rg.resize(i); }
    inline void push_back(void* const pItem) { m_rg.push_back(pItem); }
 
    inline int find(const void * const pvItem) const
